Move prompt to its own file
